+++ CHANGELOG.md
@@ -1,4 +1,2 @@
 # What's new:
-- Improved the API for developers to use
-- Fixed some heads/skulls having a hole in them
-- Updated CarvedWood compatibility to fix a crash on latest update
+- Added full VulkanMod compatibility